Use u32 for indexes and sizes rather than usize

+++ crates/js-sys/src/lib.rs
@@ -45,6 +45,8 @@ use wasm_bindgen::JsCast;
 //
 // * Arguments that are `JsValue`s or imported JavaScript types should be taken
 //   by reference.
+//
+// * Prefer using `u32` for indexes and sizes rather than `usize`.
 
 #[wasm_bindgen]
 extern "C" {
@@ -498,7 +500,7 @@ extern "C" {
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView)
     #[wasm_bindgen(constructor)]
-    pub fn new(buffer: &ArrayBuffer, byteOffset: usize, byteLength: usize) -> DataView;
+    pub fn new(buffer: &ArrayBuffer, byteOffset: u32, byteLength: u32) -> DataView;
 
     /// The ArrayBuffer referenced by this view. Fixed at construction time and thus read only.
     ///
@@ -511,210 +513,210 @@ extern "C" {
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/byteLength)
     #[wasm_bindgen(method, getter, structural, js_name = byteLength)]
-    pub fn byte_length(this: &DataView) -> usize;
+    pub fn byte_length(this: &DataView) -> u32;
 
     /// The offset (in bytes) of this view from the start of its ArrayBuffer.
     /// Fixed at construction time and thus read only.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/byteOffset)
     #[wasm_bindgen(method, getter, structural, js_name = byteOffset)]
-    pub fn byte_offset(this: &DataView) -> usize;
+    pub fn byte_offset(this: &DataView) -> u32;
 
     /// The getInt8() method gets a signed 8-bit integer (byte) at the
     /// specified byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getInt8)
     #[wasm_bindgen(method, js_name = getInt8)]
-    pub fn get_int8(this: &DataView, byte_offset: usize) -> i8;
+    pub fn get_int8(this: &DataView, byte_offset: u32) -> i8;
 
     /// The getUint8() method gets a unsigned 8-bit integer (byte) at the specified
     /// byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getUint8)
     #[wasm_bindgen(method, js_name = getUint8)]
-    pub fn get_uint8(this: &DataView, byte_offset: usize) -> u8;
+    pub fn get_uint8(this: &DataView, byte_offset: u32) -> u8;
 
     /// The getInt16() method gets a signed 16-bit integer (byte) at the specified
     /// byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getInt16)
     #[wasm_bindgen(method, js_name = getInt16)]
-    pub fn get_int16(this: &DataView, byte_offset: usize) -> i16;
+    pub fn get_int16(this: &DataView, byte_offset: u32) -> i16;
 
     /// The getInt16() method gets a signed 16-bit integer (byte) at the specified
     /// byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getInt16)
     #[wasm_bindgen(method, js_name = getInt16)]
-    pub fn get_int16_endian(this: &DataView, byte_offset: usize, little_endian: bool) -> i16;
+    pub fn get_int16_endian(this: &DataView, byte_offset: u32, little_endian: bool) -> i16;
 
     /// The getUint16() an unsigned 16-bit integer (unsigned byte) at the specified
     /// byte offset from the start of the view.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getUint16)
     #[wasm_bindgen(method, js_name = getUint16)]
-    pub fn get_uint16(this: &DataView, byte_offset: usize) -> u16;
+    pub fn get_uint16(this: &DataView, byte_offset: u32) -> u16;
 
     /// The getUint16() an unsigned 16-bit integer (unsigned byte) at the specified
     /// byte offset from the start of the view.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getUint16)
     #[wasm_bindgen(method, js_name = getUint16)]
-    pub fn get_uint16_endian(this: &DataView, byte_offset: usize, little_endian: bool) -> u16;
+    pub fn get_uint16_endian(this: &DataView, byte_offset: u32, little_endian: bool) -> u16;
 
     /// The getInt32() method gets a signed 16-bit integer (byte) at the specified
     /// byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getInt32)
     #[wasm_bindgen(method, js_name = getInt32)]
-    pub fn get_int32(this: &DataView, byte_offset: usize) -> i32;
+    pub fn get_int32(this: &DataView, byte_offset: u32) -> i32;
 
     /// The getInt32() method gets a signed 16-bit integer (byte) at the specified
     /// byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getInt32)
     #[wasm_bindgen(method, js_name = getInt32)]
-    pub fn get_int32_endian(this: &DataView, byte_offset: usize, little_endian: bool) -> i32;
+    pub fn get_int32_endian(this: &DataView, byte_offset: u32, little_endian: bool) -> i32;
 
     /// The getUint32() an unsigned 16-bit integer (unsigned byte) at the specified
     /// byte offset from the start of the view.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getUint32)
     #[wasm_bindgen(method, js_name = getUint32)]
-    pub fn get_uint32(this: &DataView, byte_offset: usize) -> u32;
+    pub fn get_uint32(this: &DataView, byte_offset: u32) -> u32;
 
     /// The getUint32() an unsigned 16-bit integer (unsigned byte) at the specified
     /// byte offset from the start of the view.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getUint32)
     #[wasm_bindgen(method, js_name = getUint32)]
-    pub fn get_uint32_endian(this: &DataView, byte_offset: usize, little_endian: bool) -> u32;
+    pub fn get_uint32_endian(this: &DataView, byte_offset: u32, little_endian: bool) -> u32;
 
     /// The getFloat32() method gets a signed 32-bit float (float) at the specified
     /// byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getFloat32)
     #[wasm_bindgen(method, js_name = getFloat32)]
-    pub fn get_float32(this: &DataView, byte_offset: usize) -> f32;
+    pub fn get_float32(this: &DataView, byte_offset: u32) -> f32;
 
     /// The getFloat32() method gets a signed 32-bit float (float) at the specified
     /// byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getFloat32)
     #[wasm_bindgen(method, js_name = getFloat32)]
-    pub fn get_float32_endian(this: &DataView, byte_offset: usize, little_endian: bool) -> f32;
+    pub fn get_float32_endian(this: &DataView, byte_offset: u32, little_endian: bool) -> f32;
 
     /// The getFloat64() method gets a signed 64-bit float (float) at the specified
     /// byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getFloat64)
     #[wasm_bindgen(method, js_name = getFloat64)]
-    pub fn get_float64(this: &DataView, byte_offset: usize) -> f64;
+    pub fn get_float64(this: &DataView, byte_offset: u32) -> f64;
 
     /// The getFloat64() method gets a signed 64-bit float (float) at the specified
     /// byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getFloat64)
     #[wasm_bindgen(method, js_name = getFloat64)]
-    pub fn get_float64_endian(this: &DataView, byte_offset: usize, little_endian: bool) -> f64;
+    pub fn get_float64_endian(this: &DataView, byte_offset: u32, little_endian: bool) -> f64;
 
     /// The setInt8() method stores a signed 8-bit integer (byte) value at the
     /// specified byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setInt8)
     #[wasm_bindgen(method, js_name = setInt8)]
-    pub fn set_int8(this: &DataView, byte_offset: usize, value: i8);
+    pub fn set_int8(this: &DataView, byte_offset: u32, value: i8);
 
     /// The setUint8() method stores an unsigned 8-bit integer (byte) value at the
     /// specified byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setUint8)
     #[wasm_bindgen(method, js_name = setUint8)]
-    pub fn set_uint8(this: &DataView, byte_offset: usize, value: u8);
+    pub fn set_uint8(this: &DataView, byte_offset: u32, value: u8);
 
     /// The setInt16() method stores a signed 16-bit integer (byte) value at the
     /// specified byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setInt16)
     #[wasm_bindgen(method, js_name = setInt16)]
-    pub fn set_int16(this: &DataView, byte_offset: usize, value: i16);
+    pub fn set_int16(this: &DataView, byte_offset: u32, value: i16);
 
     /// The setInt16() method stores a signed 16-bit integer (byte) value at the
     /// specified byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setInt16)
     #[wasm_bindgen(method, js_name = setInt16)]
-    pub fn set_int16_endian(this: &DataView, byte_offset: usize, value: i16, little_endian: bool);
+    pub fn set_int16_endian(this: &DataView, byte_offset: u32, value: i16, little_endian: bool);
 
     /// The setUint16() method stores an unsigned 16-bit integer (byte) value at the
     /// specified byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setUint16)
     #[wasm_bindgen(method, js_name = setUint16)]
-    pub fn set_uint16(this: &DataView, byte_offset: usize, value: u16);
+    pub fn set_uint16(this: &DataView, byte_offset: u32, value: u16);
 
     /// The setUint16() method stores an unsigned 16-bit integer (byte) value at the
     /// specified byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setUint16)
     #[wasm_bindgen(method, js_name = setUint16)]
-    pub fn set_uint16_endian(this: &DataView, byte_offset: usize, value: u16, little_endian: bool);
+    pub fn set_uint16_endian(this: &DataView, byte_offset: u32, value: u16, little_endian: bool);
 
     /// The setInt32() method stores a signed 32-bit integer (byte) value at the
     /// specified byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setInt32)
     #[wasm_bindgen(method, js_name = setInt32)]
-    pub fn set_int32(this: &DataView, byte_offset: usize, value: i32);
+    pub fn set_int32(this: &DataView, byte_offset: u32, value: i32);
 
     /// The setInt32() method stores a signed 32-bit integer (byte) value at the
     /// specified byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setInt32)
     #[wasm_bindgen(method, js_name = setInt32)]
-    pub fn set_int32_endian(this: &DataView, byte_offset: usize, value: i32, little_endian: bool);
+    pub fn set_int32_endian(this: &DataView, byte_offset: u32, value: i32, little_endian: bool);
 
     /// The setUint32() method stores an unsigned 32-bit integer (byte) value at the
     /// specified byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setUint32)
     #[wasm_bindgen(method, js_name = setUint32)]
-    pub fn set_uint32(this: &DataView, byte_offset: usize, value: u32);
+    pub fn set_uint32(this: &DataView, byte_offset: u32, value: u32);
 
     /// The setUint32() method stores an unsigned 32-bit integer (byte) value at the
     /// specified byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setUint32)
     #[wasm_bindgen(method, js_name = setUint32)]
-    pub fn set_uint32_endian(this: &DataView, byte_offset: usize, value: u32, little_endian: bool);
+    pub fn set_uint32_endian(this: &DataView, byte_offset: u32, value: u32, little_endian: bool);
 
     /// The setFloat32() method stores a signed 32-bit float (float) value at the
     /// specified byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setFloat32)
     #[wasm_bindgen(method, js_name = setFloat32)]
-    pub fn set_float32(this: &DataView, byte_offset: usize, value: f32);
+    pub fn set_float32(this: &DataView, byte_offset: u32, value: f32);
 
     /// The setFloat32() method stores a signed 32-bit float (float) value at the
     /// specified byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setFloat32)
     #[wasm_bindgen(method, js_name = setFloat32)]
-    pub fn set_float32_endian(this: &DataView, byte_offset: usize, value: f32, little_endian: bool);
+    pub fn set_float32_endian(this: &DataView, byte_offset: u32, value: f32, little_endian: bool);
 
     /// The setFloat64() method stores a signed 64-bit float (float) value at the
     /// specified byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setFloat64)
     #[wasm_bindgen(method, js_name = setFloat64)]
-    pub fn set_float64(this: &DataView, byte_offset: usize, value: f64);
+    pub fn set_float64(this: &DataView, byte_offset: u32, value: f64);
 
     /// The setFloat64() method stores a signed 64-bit float (float) value at the
     /// specified byte offset from the start of the DataView.
     ///
     /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setFloat64)
     #[wasm_bindgen(method, js_name = setFloat64)]
-    pub fn set_float64_endian(this: &DataView, byte_offset: usize, value: f64, little_endian: bool);
+    pub fn set_float64_endian(this: &DataView, byte_offset: u32, value: f64, little_endian: bool);
 }
 
 // Error
